# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talim AI - Backend для тестирования компетенций",
    "description": "\n    API для системы AI-тестирования профессиональных компетенций через голосовое собеседование.\n    \n    ## Основные возможности\n    \n    * **Тестирование компетенций**: Адаптивное голосовое тестирование с использованием GPT-4\n    * **Транскрипция аудио**: Автоматическая транскрипция ответов через Whisper API\n    * **Оценка ответов**: AI-оценка с выявлением пробелов в знаниях\n    \n    ## Документация\n    \n    * Swagger UI: `/docs`\n    * ReDoc: `/redoc`\n    * OpenAPI JSON: `/openapi.json`\n    \n    ## Аутентификация\n    \n    User ID передается с фронтенда в заголовке `Authorization`:\n    - Формат: `Authorization: Bearer {user_id}` или `Authorization: {user_id}`\n    - Заголовок обязателен для всех защищенных endpoints\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/questions/generate": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Получить вопрос",
        "description": "Получает вопрос для тестирования компетенции из базы данных. Вопрос должен быть предварительно добавлен в БД.",
        "operationId": "generate_question_api_questions_generate_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_question_api_questions_generate_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/answer": {
      "post": {
        "tags": [
          "questions"
        ],
        "summary": "Отправить голосовой ответ",
        "description": "Отправляет голосовой ответ на вопрос. Аудио файл транскрибируется через Whisper API, после чего ответ оценивается через GPT-4. Вопрос и ответ сохраняются в БД только после отправки.",
        "operationId": "submit_answer_api_questions_answer_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене.",
              "title": "Authorization"
            },
            "description": "JWT токен передается с фронтенда в заголовке Authorization. Формат: 'Bearer {jwt_token}'. User ID извлекается из поля 'sub' в JWT токене."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_answer_api_questions_answer_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerEvaluation": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "understandingDepth": {
            "type": "string",
            "title": "Understandingdepth"
          },
          "isCorrect": {
            "type": "boolean",
            "title": "Iscorrect"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "knowledgeGaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledgegaps"
          },
          "nextDifficulty": {
            "type": "integer",
            "title": "Nextdifficulty"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "correctAnswer": {
            "type": "string",
            "title": "Correctanswer"
          },
          "expectedKeyPoints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expectedkeypoints"
          }
        },
        "type": "object",
        "required": [
          "score",
          "understandingDepth",
          "isCorrect",
          "feedback",
          "knowledgeGaps",
          "nextDifficulty",
          "correctAnswer",
          "expectedKeyPoints"
        ],
        "title": "AnswerEvaluation",
        "description": "Оценка ответа"
      },
      "AnswerResponse": {
        "properties": {
          "transcript": {
            "type": "string",
            "title": "Transcript"
          },
          "evaluation": {
            "$ref": "#/components/schemas/AnswerEvaluation"
          }
        },
        "type": "object",
        "required": [
          "transcript",
          "evaluation"
        ],
        "title": "AnswerResponse",
        "description": "Ответ на отправку голосового ответа"
      },
...
      "Body_generate_question_api_questions_generate_post": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assessment Id",
            "description": "ID тестирования"
          },
          "competency_id": {
            "type": "string",
            "format": "uuid",
            "title": "Competency Id",
            "description": "ID компетенции"
          },
          "question_number": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Question Number",
            "description": "Номер вопроса (1-5)"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty",
            "description": "Уровень сложности (1-5). Если не указан, определяется автоматически"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "competency_id",
          "question_number"
        ],
        "title": "Body_generate_question_api_questions_generate_post"
      },
      "Body_submit_answer_api_questions_answer_post": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assessment Id",
            "description": "ID тестирования"
          },
          "competency_id": {
            "type": "string",
            "format": "uuid",
            "title": "Competency Id",
            "description": "ID компетенции"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text",
            "description": "Текст вопроса, на который отвечает пользователь"
          },
          "difficulty": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty",
            "description": "Сложность вопроса (1-5)",
            "default": 3
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id",
            "description": "ID вопроса из БД (если был использован сохраненный вопрос)"
          },
          "audio": {
            "type": "string",
            "format": "binary",
            "title": "Audio",
            "description": "Аудио файл с ответом (webm, mp3, wav, m4a, ogg). Максимум 25MB"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "competency_id",
          "question_text",
          "audio"
        ],
        "title": "Body_submit_answer_api_questions_answer_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QuestionGenerateResponse": {
        "properties": {
          "questionId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questionid"
          },
          "questionText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questiontext"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "estimatedAnswerTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimatedanswertime"
          },
          "expectedKeyPoints": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedkeypoints"
          },
          "noMoreQuestions": {
            "type": "boolean",
            "title": "Nomorequestions",
            "default": false
          }
        },
        "type": "object",
        "title": "QuestionGenerateResponse",
        "description": "Ответ на генерацию вопроса"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}